# golang_updates/drt-repos/drt-go-chain-proxy-1.2.0/cmd/proxy/config/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dharitri Gateway API",
    "description": "![Dharitri](/mvx-icon.png)\n\nThe documentation describes the endpoints that are available on the drt-go-chain-proxy project. \n\nGithub URL: `https://github.com/TerraDharitri/drt-go-chain-proxy`\n\nThis is used on `gateway.dharitri.org` while the public api (`api.dharitri.org`) is a wrapper over it.\n\nThis API is organized around REST principles, so if you've interacted with RESTful APIs before, many of the concepts will look familiar.\n\n## Request / Response Format\n\nJSON will be returned for all responses, including errors. Empty or blank fields are omitted. Requests with a message body use JSON as well. Successful requests will return a `2xx` HTTP status.\n\nThe general structure of a response is:\n``` JSON\n{\n  data  anything\n  error string\n  code  string\n}\n```\nWhere:\n- data: the expected result if the request was successful. `null` otherwise\n- error: if the request failed, the reason will be returned in this field. empty otherwise\n- code: the internal code for the request. can be `successful`, `internal_issue` or `bad_request`\n",
    "version": "1.1.0"
  },
  "paths": {
...
    "/block/{shard}/by-nonce/{nonce}": {
      "get": {
        "tags": [
          "block"
        ],
        "summary": "return a block by nonce and shard.",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "nonce",
            "in": "path",
            "description": "the nonce to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/block/{shard}/by-nonce/{nonce}?withTxs=true&withLogs=true&forHyperblock=true": {
      "get": {
        "tags": [
          "block"
        ],
        "summary": "returns a block by nonce, with transactions included.",
        "parameters": [
          {
            "name": "nonce",
            "in": "path",
            "description": "the nonce to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to query the block for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/block/{shard}/by-hash/{hash}": {
      "get": {
        "tags": [
          "block"
        ],
        "summary": "returns a block by hash.",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "the hash to look after",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to query the block for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/block/{shard}/by-hash/{hash}?withTxs=true&withLogs=true&forHyperblock=true": {
      "get": {
        "tags": [
          "block"
        ],
        "summary": "returns a block by hash, with transactions included.",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "the hash to look after",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to query the block for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/hyperblock/by-nonce/{nonce}": {
      "get": {
        "tags": [
          "hyperblock"
        ],
        "summary": "get the hyperblock at the specified nonce",
        "parameters": [
          {
            "name": "nonce",
            "in": "path",
            "description": "the nonce to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/hyperblock/by-hash/{hash}": {
      "get": {
        "tags": [
          "hyperblock"
        ],
        "summary": "get the hyperblock with the specific hash",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "the hash to look after",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/{shard}/raw/block/by-nonce/{nonce}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the block at the given nonce following the protocol structure",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "nonce",
            "in": "path",
            "description": "the nonce to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/{shard}/raw/block/by-hash/{hash}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the block with the given hash following the protocol structure",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "description": "the hash to look after",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/{shard}/json/block/by-nonce/{nonce}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the block at the given nonce following the protocol structure, JSON encoded",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "nonce",
            "in": "path",
            "description": "the nonce to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
    "/internal/{shard}/json/block/by-hash/{hash}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "returns the block with the given hash following the protocol structure, JSON encoded",
        "parameters": [
          {
            "name": "shard",
            "in": "path",
            "description": "the shard ID to look after",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "description": "the hash to look after",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          }
        }
      }
    },
...
  },
...
  "servers": [],
  "components": {
...
    "schemas": {
      "GenericResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
    }
  }
}